pyproject.toml
# PEP 518: https://www.python.org/dev/peps/pep-0518/

[tool.black]
line-length = 150

[tool.isort]
force_single_line = true
known_first_party = "grouprep"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
]

[tool.poetry]
name = "grouprep"
version = "1.0.0"
description = "Sambi Group Representative Information Exchange MD tool"
authors = ["Stefan Halén <stefan.halen@internetstiftelsen.se>"]
license = "Apache-2.0"
repository = "https://github.com/s-hal/grouprep"
readme = "README.md"
packages = [
    { include = "grouprep", from = "src" }
]

[tool.poetry.scripts]
grsign = "grouprep.grsign:main"
grverify = "grouprep.grverify:main"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
pyyaml = "^6.0"
cryptojwt = "^1.8.3"
jsonschema = "^4.19.0"
jose = "^1.0.0"


[tool.poetry.group.dev.dependencies]
responses = "^0.23.1"
pytest-cov = "^4.1.0"
alabaster = "^0.7.12"
black = "^22"
isort = "^5.6.4"
pytest = "^6.1.2"
pytest-mock = "^3.11.1"
pytest-black = "^0.3.12"
pytest-isort = "^1.2.0"
sphinx = "^3.5.2"
sphinx-autobuild = "^2021.3.14"
#codecov = "^2.1.11"
coverage = "^6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
